pyproject.toml
[project]
name = "hackathonmirrorproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "llama-index-core>=0.14.12",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-llms-openai>=0.6.13",
    "openai>=2.15.0",
    "openai-agents>=0.6.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "ruff>=0.14.0",
]


[tool.ruff]
target-version = "py312"

exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "B",
    "UP",
]

ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["mirror"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
